requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
dj-database-url==1.0.0
Django==4.0.6
django-formtools==2.3
django-heroku==0.3.1
django-rest-framework==0.1.0
django-xhtml2pdf==0.0.4
djangorestframework==3.13.1
future==0.18.2
geodepy==0.4.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
joblib==1.1.0
lxml==4.9.1
numpy==1.23.1
oscrypto==1.3.0
pandas==1.4.3
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyPDF2==2.10.3
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.6
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.10
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.5
